# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "High-performance core for VELTRA NFC routing, Ledgers, and AI banking.",
        "title": "VELTRA API Engine",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/v1/ledger/reconcile": {
            "post": {
                "description": "Accepts a cryptographically signed receipt from the offline wallet to settle funds online.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ledger"
                ],
                "summary": "Reconcile Offline Transaction",
                "parameters": [
                    {
                        "description": "Offline Receipt Details",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_ledger.OfflineReceipt"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Settle successful",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/merchant/expenses": {
            "post": {
                "description": "Records a new expense entry (Raw Materials, Logistics, etc).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant"
                ],
                "summary": "Log Business Expense",
                "parameters": [
                    {
                        "description": "Expense Details",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_ledger.Expense"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Expense logged",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/merchant/inventory": {
            "get": {
                "description": "Retrieves all items in stock for the merchant.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant"
                ],
                "summary": "Get Merchant Inventory",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_ledger.InventoryItem"
                            }
                        }
                    }
                }
            }
        },
        "/v1/payments/nfc-tap": {
            "post": {
                "description": "Accepts encrypted card/phone tap data and passes it through the ACID transaction ledger.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Process NFC Tap Payment",
                "parameters": [
                    {
                        "description": "NFC Tap Transaction Details",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_ledger.NFCPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_ledger.PaymentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid Payload Structure",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "402": {
                        "description": "Payment Required - Insufficient Balance",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/payments/secure-nfc-tap": {
            "post": {
                "description": "Decrypts AES-256-GCM data and passes it through the ACID transaction ledger.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Process Secure NFC Tap Payment",
                "parameters": [
                    {
                        "description": "Encrypted NFC Data",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_ledger.SecureNFCPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_ledger.PaymentResponse"
                        }
                    }
                }
            }
        },
        "/v1/pockets": {
            "get": {
                "description": "Retrieves all savings, squad, and offline pockets for the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pockets"
                ],
                "summary": "List User Pockets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_ledger.Pocket"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new savings or squad pocket.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pockets"
                ],
                "summary": "Create New Pocket",
                "parameters": [
                    {
                        "description": "Pocket Details",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_ledger.Pocket"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Pocket created",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_ledger.Expense": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "merchant_id": {
                    "type": "string"
                }
            }
        },
        "internal_ledger.InventoryItem": {
            "type": "object",
            "properties": {
                "cost_price": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "merchant_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "selling_price": {
                    "type": "integer"
                },
                "stock_count": {
                    "type": "integer"
                }
            }
        },
        "internal_ledger.NFCPayload": {
            "type": "object",
            "required": [
                "amount",
                "device_signature",
                "sender_id"
            ],
            "properties": {
                "amount": {
                    "description": "5,000.00 NGN in Kobo",
                    "type": "integer",
                    "example": 500000
                },
                "device_signature": {
                    "type": "string",
                    "example": "a4f2e9..."
                },
                "receiver_id": {
                    "type": "string",
                    "example": "merch_1122"
                },
                "sender_id": {
                    "type": "string",
                    "example": "usr_99823"
                }
            }
        },
        "internal_ledger.OfflineReceipt": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "cert_id": {
                    "type": "string"
                },
                "device_sig": {
                    "type": "string"
                },
                "hardware_counter": {
                    "type": "integer"
                },
                "merchant_id": {
                    "type": "string"
                },
                "prev_hash": {
                    "type": "string"
                },
                "tx_ref": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "internal_ledger.PaymentResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Transfer successful"
                },
                "status": {
                    "type": "string",
                    "example": "APPROVED"
                },
                "transaction_id": {
                    "type": "string",
                    "example": "tx_7718293"
                }
            }
        },
        "internal_ledger.Pocket": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "target_amount": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "internal_ledger.SecureNFCPayload": {
            "type": "object",
            "required": [
                "encrypted_data"
            ],
            "properties": {
                "encrypted_data": {
                    "type": "string"
                }
            }
        }
    }
}